+++ docs/Whats-New.md
@@ -53,6 +53,9 @@ This page lists the history of changes across stable Vinifera releases and also 
 
 :::{dropdown} Click to show
 
+New:
+- Implement voxel light customization (by ZivDero)
+
 Vinifera fixes:
 - Fix unit placement in non-TS Client builds of Vinifera (by ZivDero)